+++ test/runtime/samples/bitmask-overflow-if/_config.js
@@ -1,0 +1,24 @@
+export default {
+	html: `
+		012345678910111213141516171819202122232425262728293031323334353637383940
+		expected: true
+		if: true
+		<button></button>
+	`,
+
+	async test({ assert, component, target, window }) {
+		const button = target.querySelector("button");
+		await button.dispatchEvent(new window.MouseEvent("click"));
+
+		assert.htmlEqual(
+			target.innerHTML,
+			`
+			112345678910111213141516171819202122232425262728293031323334353637383940
+			expected: false
+			if: false
+			<div></div>
+			<button></button>
+		`
+		);
+	}
+};